remove batch syncer and move the unit test utils
